+++ changelog.md
@@ -591,3 +591,9 @@
         - Fixed an issue where random chance upgrades would provide inconsistent values when calculating coin projections
             - The coin projection (i.e "Your factory will produce x - x amount of coins per process cycle") should now be 100% accurate!
         - Fixed an issue where the Fully Automated Workforce upgrade could cause a coin wipe during coin projection calculation, causing incorrection projection values
+
+2.902
+    Coin Factory
+        - Added the "Purchase All Upgrades" button to the homepage of the wizard
+            - This button will replace the "prestige" button if all upgrades are not currently owned.
+            - This button will purchase all upgrades that can possibly be purchased with the factory's invested coins.